requirements.txt
langchain==0.0.208
openai==0.27.8
chromadb==0.3.26
tqdm==4.65.0
python-dotenv==1.0.0
unstructured==0.6.7
sentence-transformers==2.2.2
PyPDF2==3.0.1
